app/golf_project/requirements.txt
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.4
django-filter==1.1.0
django-toolbelt==0.0.1
djangorestframework==3.8.0
gunicorn==19.7.1
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.2
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
scipy==1.0.1
seaborn==0.8.1
six==1.11.0
static3==0.7.0
